+++ app/Actions/Post/UpdatePost.php
@@ -9,6 +9,7 @@ use App\Enums\Post\Status as PostStatus;
 use App\Jobs\PublishPost;
 use App\Models\Post;
 use App\Models\Workspace;
+use App\Support\PostStatusRules;
 use Carbon\Carbon;
 use Illuminate\Support\Arr;
 use Illuminate\Support\Facades\DB;
@@ -20,7 +21,7 @@ class UpdatePost
      */
     public static function execute(Workspace $workspace, Post $post, array $data): array
     {
-        if (PostStatusGuard::blocksEditing($post)) {
+        if (PostStatusRules::blocksEditing($post)) {
             return ['post' => $post, 'action' => PostAction::Finalized];
         }
