+++ app/Http/Controllers/Api/PostController.php
@@ -6,7 +6,6 @@ namespace App\Http\Controllers\Api;
 
 use App\Actions\Post\CreatePost;
 use App\Actions\Post\DeletePost;
-use App\Actions\Post\PostStatusGuard;
 use App\Actions\Post\UpdatePost;
 use App\Enums\Media\Type as MediaType;
 use App\Enums\Post\Action as PostAction;
@@ -20,6 +19,7 @@ use App\Http\Resources\Api\PostPreviewResource;
 use App\Http\Resources\Api\PostResource;
 use App\Models\Post;
 use App\Services\Post\MediaAttacher;
+use App\Support\PostStatusRules;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
 use Illuminate\Http\Resources\Json\AnonymousResourceCollection;
@@ -70,7 +70,7 @@ class PostController extends Controller
 
         if (data_get($result, 'action') === PostAction::Finalized) {
             return response()->json(
-                ['message' => PostStatusGuard::editBlockedMessage()],
+                ['message' => PostStatusRules::editBlockedMessage()],
                 Response::HTTP_UNPROCESSABLE_ENTITY
             );
         }
