+++ tests/Unit/Support/PostStatusRulesTest.php
@@ -2,14 +2,14 @@
 
 declare(strict_types=1);
 
-use App\Actions\Post\PostStatusGuard;
 use App\Enums\Post\Status as PostStatus;
 use App\Models\Post;
+use App\Support\PostStatusRules;
 
 test('blocks editing for terminal statuses', function (PostStatus $status) {
     $post = Post::factory()->make(['status' => $status]);
 
-    expect(PostStatusGuard::blocksEditing($post))->toBeTrue();
+    expect(PostStatusRules::blocksEditing($post))->toBeTrue();
 })->with([
     PostStatus::Publishing,
     PostStatus::Published,
@@ -20,7 +20,7 @@ test('blocks editing for terminal statuses', function (PostStatus $status) {
 test('allows editing for non terminal statuses', function (PostStatus $status) {
     $post = Post::factory()->make(['status' => $status]);
 
-    expect(PostStatusGuard::blocksEditing($post))->toBeFalse();
+    expect(PostStatusRules::blocksEditing($post))->toBeFalse();
 })->with([
     PostStatus::Draft,
     PostStatus::Scheduled,
@@ -29,7 +29,7 @@ test('allows editing for non terminal statuses', function (PostStatus $status) {
 test('blocks deletion for published statuses', function (PostStatus $status) {
     $post = Post::factory()->make(['status' => $status]);
 
-    expect(PostStatusGuard::blocksDeletion($post))->toBeTrue();
+    expect(PostStatusRules::blocksDeletion($post))->toBeTrue();
 })->with([
     PostStatus::Publishing,
     PostStatus::Published,
@@ -39,7 +39,7 @@ test('blocks deletion for published statuses', function (PostStatus $status) {
 test('allows deletion for draft, scheduled and failed statuses', function (PostStatus $status) {
     $post = Post::factory()->make(['status' => $status]);
 
-    expect(PostStatusGuard::blocksDeletion($post))->toBeFalse();
+    expect(PostStatusRules::blocksDeletion($post))->toBeFalse();
 })->with([
     PostStatus::Draft,
     PostStatus::Scheduled,
